DeadSystem+Dame theo thoi gian
ok

+++ CopyNgoiSaoBolac/Assets/Scripts/WorldSpaceUIController.cs
@@ -31,7 +31,7 @@ public class WorldSpaceUIController : MonoBehaviour
         dealDamageSystem.OnGrantExperience -= DisplayExperienceIcon;
     }
 
-    private void DisplayDamageIcon(int damageAmount, float3 startPosition)
+    private void DisplayDamageIcon(float damageAmount, float3 startPosition)
     {
         var directionToCamera = (Vector3)startPosition - _mainCameraTransform.position;
         var rotationToCamera = Quaternion.LookRotation(directionToCamera, Vector3.up);
@@ -40,7 +40,7 @@ public class WorldSpaceUIController : MonoBehaviour
         newIconText.text = $"<color=red>-{damageAmount.ToString()}</color>";
     }
 
-    private void DisplayExperienceIcon(int experienceAmount, float3 startPosition)
+    private void DisplayExperienceIcon(float experienceAmount, float3 startPosition)
     {
         var directionToCamera = (Vector3)startPosition - _mainCameraTransform.position;
         var rotationToCamera = Quaternion.LookRotation(directionToCamera, Vector3.up);
